+++ src/context/subscription_task.rs
@@ -1,5 +1,5 @@
 use crate::{
-    context::{LongPoolingServiceContext, Subscription},
+    context::{Channel, LongPoolingServiceContext},
     messages::SubscriptionMessage,
 };
 use serde_json::Value as JsonValue;
@@ -7,37 +7,37 @@ use std::sync::Arc;
 use tokio::sync::mpsc;
 
 pub(crate) fn spawn(
-    subscription: String,
+    channel: String,
     mut rx: mpsc::Receiver<JsonValue>,
     inner: Arc<LongPoolingServiceContext>,
 ) {
     tokio::task::spawn(async move {
         while let Some(msg) = rx.recv().await {
             tracing::debug!(
-                subscription = subscription,
-                "`{subscription}` channel got message: `{msg:?}`."
+                channel = channel,
+                "`{channel}` channel got message: `{msg:?}`."
             );
 
-            let client_id_channels = inner.client_id_channels.read().await;
+            let client_id_channels = inner.client_id_senders.read().await;
 
             for (client_id, client_channel) in inner
-                .subscriptions_data
+                .channels_data
                 .read()
                 .await
-                .get(&subscription)
+                .get(&channel)
                 .into_iter()
-                .flat_map(Subscription::client_ids)
+                .flat_map(Channel::client_ids)
                 .filter_map(|client_id| client_id_channels.get(client_id).map(|v| (client_id, v)))
             {
                 tracing::trace!(
                     client_id = %client_id,
-                    subscription = subscription,
-                    "Message `{msg:?}` from channel `{subscription}` was sent to client `{client_id}`."
+                    channel = channel,
+                    "Message `{msg:?}` from channel `{channel}` was sent to client `{client_id}`."
                 );
 
                 if client_channel
                     .send(SubscriptionMessage {
-                        subscription: subscription.clone(),
+                        subscription: channel.clone(),
                         msg: msg.clone(),
                     })
                     .await
@@ -45,7 +45,7 @@ pub(crate) fn spawn(
                 {
                     tracing::error!(
                         client_id = %client_id,
-                        subscription = subscription,
+                        channel = channel,
                         "Channel was closed!"
                     );
                 }