+++ src/handlers/connect/publish.rs
@@ -1,4 +1,5 @@
-use crate::{context::Subscription, messages::Message, LongPoolingServiceContext};
+use crate::messages::Advice;
+use crate::{context::Channel, messages::Message, LongPoolingServiceContext};
 use axum::http::StatusCode;
 
 #[inline]
@@ -7,11 +8,11 @@ pub(super) async fn publish_handle(
     mut messages: Vec<Message>,
 ) -> Result<Vec<Message>, StatusCode> {
     if messages.iter().any(|message| {
-        if let Some(channel) = &message.channel {
-            channel.contains("/meta/")
-        } else {
-            true
-        }
+        message
+            .channel
+            .as_ref()
+            .map(|channel| channel.contains("/meta/") && channel != "/meta/connect")
+            .unwrap_or(false)
     }) {
         Err(StatusCode::BAD_REQUEST)
     } else {
@@ -26,39 +27,37 @@ pub(super) async fn publish_handle(
                 ..
             } = std::mem::take(message);
 
-            *message = match (channel, data, client_id) {
-                (None, _, _) => Message::error("400::channel_missing", None, None, id),
-                (_, None, _) => Message::error("400::data_missing", None, None, id),
-                (_, _, None) => Message::error("400::client_id_missing", None, None, id),
-                (Some(subscription), Some(data), Some(client_id)) => {
+            *message = match (channel, client_id) {
+                (None, _) => Message::channel_missing(id),
+                (channel, None) => Message::session_unknown(id, channel, Some(Advice::handshake())),
+                (Some(channel), Some(client_id)) => {
                     if context.check_client_id(&client_id).await {
-                        if let Some(tx) = subscriptions_data_read_guard
-                            .get(&subscription)
-                            .map(Subscription::tx)
+                        if let Some(tx) =
+                            subscriptions_data_read_guard.get(&channel).map(Channel::tx)
                         {
-                            if tx.send(data).await.is_err() {
+                            if tx.send(data.unwrap_or_default()).await.is_err() {
                                 tracing::error!(
                                     client_id = %client_id,
-                                    subscription = subscription,
+                                    subscription = channel,
                                     "Channel was closed!"
                                 );
                             }
                         } else {
                             tracing::trace!(
                                 client_id = %client_id,
-                                subscription = subscription,
-                                "No `{subscription}` channel was found for message: `{data:?}`."
+                                channel = channel,
+                                "No `{channel}` channel was found for message: `{data:?}`."
                             );
                         }
 
                         Message {
                             id,
-                            channel: Some(subscription),
+                            channel: Some(channel),
                             successful: Some(true),
                             ..Default::default()
                         }
                     } else {
-                        Message::error("402::session_unknown", Some(subscription), None, id)
+                        Message::session_unknown(id, Some(channel), None)
                     }
                 }
             };