+++ src/types/subscription_id.rs
@@ -1,2 +1,2 @@
 // TODO: Replace on Arc<str>?
-pub(crate) type SubscriptionId = String;
+pub(crate) type ChannelId = String;